requirements.txt
setuptools
shapely
discord.py
scikit-image
paddleocr
paddlepaddle
PyAutoGUI
PyGetWindow
langchain
langchain-groq
pandas
imgaug
pyclipper
lmdb
tqdm
numpy<2.0
rapidfuzz
opencv-python
opencv-contrib-python
cython
Pillow
pyyaml
requests
albumentations==1.4.10
# to be compatible with albumentations
albucore==0.0.13
